Define pid_t to be an unsigned long on Windows.
This matches the actual DWORD return type of GetProcessId(). See:
https://docs.microsoft.com/en-us/windows/desktop/api/processthreadsapi/nf-processthreadsapi-getprocessid
and
https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-dtyp/262627d8-3418-4627-9218-4ffe110850b2

+++ compat/win/sys/types.h
@@ -20,6 +20,6 @@
 
 #include <stdint.h>
 
-typedef unsigned int pid_t;
+typedef unsigned long pid_t;
 
 #endif  // CRASHPAD_COMPAT_WIN_SYS_TYPES_H_
